Move uploadBufferToStorage to it's own file
