+++ functions/src/api/routes/file/files.ts
@@ -1,10 +1,7 @@
 import { FastifyInstance } from 'fastify'
 import { Static, Type } from '@sinclair/typebox'
-import { extractMultipartFormData } from './parseMultipartFiles'
-import { v4 as uuidv4 } from 'uuid'
-import firebase from 'firebase-admin'
-import { checkFileTypes } from '../../other/checkFileTypes'
-import { getStorageBucketName } from '../../dao/firebasePlugin'
+import { extractMultipartFormData } from './utils/parseMultipartFiles'
+import { uploadBufferToStorage } from './utils/uploadBufferToStorage'
 
 export const NewFile = Type.Any()
 
@@ -92,41 +89,4 @@ export const filesRoutes = (fastify: FastifyInstance, options: any, done: () => 
         }
     )
     done()
-}
-
-export const uploadBufferToStorage = async (
-    firebase: firebase.app.App,
-    buffer: Buffer,
-    eventId: string,
-    fileName: string
-): Promise<[boolean, string]> => {
-    const storageBucket = getStorageBucketName()
-
-    const fileType = await checkFileTypes(buffer, fileName)
-
-    if (!fileType) {
-        return [false, 'Invalid or unknown file type']
-    }
-
-    const { mime, extension } = fileType
-
-    const bucket = firebase.storage().bucket(storageBucket)
-    const path = `events/${eventId}/${uuidv4()}_${fileName}.${extension}`
-    const bucketFile = bucket.file(path)
-
-    try {
-        await bucketFile.save(buffer, {
-            contentType: mime,
-            predefinedAcl: 'publicRead',
-        })
-    } catch (error) {
-        console.warn('error uploading file', error)
-        const errorString = '' + error
-        return [false, 'Error uploading file, ' + errorString]
-    }
-    await bucketFile.makePublic()
-
-    const publicFileUrl = `https://${bucketFile.bucket.name}.storage.googleapis.com/${bucketFile.name}`
-
-    return [true, publicFileUrl]
 }
