+++ functions/src/api/routes/sponsors/addSponsorsPOST.ts
@@ -1,8 +1,8 @@
 import { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify'
 import { Static, Type } from '@sinclair/typebox'
 import { SponsorDao } from '../../dao/sponsorDao'
-import { uploadBufferToStorage } from '../file/files'
 import { getFileName } from '../../other/getFileName'
+import { uploadBufferToStorage } from '../file/utils/uploadBufferToStorage'
 
 export const TypeBoxSponsor = Type.Object({
     name: Type.String(),
